projects/OCR_TEXT_PROCESSING/requirements.txt
pytesseract
pillow
pandas
easyocr
opencv-python
